Change NetCDFWriter data representation
Moving to dictionary-based storage of dimensions and variables
makes it easier to access data, while also supporting more options
like writing multiple variables to the file at once.

The write method has not been updated with the new data representation,
so it will not function properly. Do not use yet.