Apply configurability to primary output
Writing of the final dataset and image files now responds to
output configuration options that can be specified at the beginning.
Accomplishing this also generalized some functions. For example, now
any kind of array-based data can be written to an image file using
ModelImageRenderer.

Not all pieces are fully implemented. For instance, no variable
attributes are written to the dataset.